fastapi/requirements.txt
PyYAML
fastapi
uvicorn
pytest
sqlmodel
jose
rsa